+++ CHANGELOG.md
@@ -28,6 +28,8 @@
 * We can add raw keyboard inputs to input editor (#326)
 * Implement ALSA poll mechanism (#321)
 * Grow the stack to its maximum at startup, to avoid unintialized stack growth (#337)
+* Add variable framerate setting (#22)
+* Encode at fixed framerate when variable framerate (#22)
 
 ### Changed